Shuffle around function call code
