+++ rust/kcl-lib/src/execution/exec_ast.rs
@@ -1,41 +1,31 @@
 use std::collections::HashMap;
 
 use async_recursion::async_recursion;
-use indexmap::IndexMap;
 
-#[cfg(feature = "artifact-graph")]
-use crate::execution::cad_op::{Group, OpArg, OpKclValue, Operation};
 use crate::{
     errors::{KclError, KclErrorDetails},
     execution::{
         annotations,
+        fn_call::Args,
         kcl_value::{FunctionSource, TypeDef},
         memory,
         state::ModuleState,
         types::{NumericType, PrimitiveType, RuntimeType},
-        BodyType, EnvironmentRef, ExecState, ExecutorContext, KclValue, Metadata, PlaneType, TagEngineInfo,
+        BodyType, EnvironmentRef, ExecState, ExecutorContext, KclValue, Metadata, PlaneType, StatementKind,
         TagIdentifier,
     },
     fmt,
     modules::{ModuleId, ModulePath, ModuleRepr},
     parsing::ast::types::{
         Annotation, ArrayExpression, ArrayRangeExpression, AscribedExpression, BinaryExpression, BinaryOperator,
-        BinaryPart, BodyItem, CallExpressionKw, Expr, FunctionExpression, IfExpression, ImportPath, ImportSelector,
-        ItemVisibility, LiteralIdentifier, LiteralValue, MemberExpression, MemberObject, Name, Node, NodeRef,
-        ObjectExpression, PipeExpression, Program, TagDeclarator, Type, UnaryExpression, UnaryOperator,
+        BinaryPart, BodyItem, Expr, IfExpression, ImportPath, ImportSelector, ItemVisibility, LiteralIdentifier,
+        LiteralValue, MemberExpression, MemberObject, Name, Node, NodeRef, ObjectExpression, PipeExpression, Program,
+        TagDeclarator, Type, UnaryExpression, UnaryOperator,
     },
     source_range::SourceRange,
-    std::{
-        args::{Arg, Args, KwArgs, TyF64},
-        FunctionKind,
-    },
+    std::args::TyF64,
     CompilationError,
 };
-
-enum StatementKind<'a> {
-    Declaration { name: &'a str },
-    Expression,
-}
 
 impl<'a> StatementKind<'a> {
     fn expect_name(&self) -> &'a str {
@@ -594,7 +584,7 @@ impl ExecutorContext {
     }
 
     #[async_recursion]
-    async fn execute_expr<'a: 'async_recursion>(
+    pub(super) async fn execute_expr<'a: 'async_recursion>(
         &self,
         init: &Expr,
         exec_state: &mut ExecState,
@@ -787,7 +777,7 @@ impl BinaryPart {
 }
 
 impl Node<Name> {
-    async fn get_result<'a>(
+    pub(super) async fn get_result<'a>(
         &self,
         exec_state: &'a mut ExecState,
         ctx: &ExecutorContext,
@@ -1305,300 +1295,6 @@ async fn inner_execute_pipe_body(
     Ok(final_output)
 }
 
-impl Node<CallExpressionKw> {
-    #[async_recursion]
-    pub async fn execute(&self, exec_state: &mut ExecState, ctx: &ExecutorContext) -> Result<KclValue, KclError> {
-        let fn_name = &self.callee;
-        let callsite: SourceRange = self.into();
-
-        // Build a hashmap from argument labels to the final evaluated values.
-        let mut fn_args = IndexMap::with_capacity(self.arguments.len());
-        let mut errors = Vec::new();
-        for arg_expr in &self.arguments {
-            let source_range = SourceRange::from(arg_expr.arg.clone());
-            let metadata = Metadata { source_range };
-            let value = ctx
-                .execute_expr(&arg_expr.arg, exec_state, &metadata, &[], StatementKind::Expression)
-                .await?;
-            let arg = Arg::new(value, source_range);
-            match &arg_expr.label {
-                Some(l) => {
-                    fn_args.insert(l.name.clone(), arg);
-                }
-                None => {
-                    if let Some(id) = arg_expr.arg.ident_name() {
-                        fn_args.insert(id.to_owned(), arg);
-                    } else {
-                        errors.push(arg);
-                    }
-                }
-            }
-        }
-
-        // Evaluate the unlabeled first param, if any exists.
-        let unlabeled = if let Some(ref arg_expr) = self.unlabeled {
-            let source_range = SourceRange::from(arg_expr.clone());
-            let metadata = Metadata { source_range };
-            let value = ctx
-                .execute_expr(arg_expr, exec_state, &metadata, &[], StatementKind::Expression)
-                .await?;
-
-            let label = arg_expr.ident_name().map(str::to_owned);
-
-            Some((label, Arg::new(value, source_range)))
-        } else {
-            None
-        };
-
-        let mut args = Args::new_kw(
-            KwArgs {
-                unlabeled,
-                labeled: fn_args,
-                errors,
-            },
-            self.into(),
-            ctx.clone(),
-            exec_state.pipe_value().map(|v| Arg::new(v.clone(), callsite)),
-        );
-        match ctx.stdlib.get_either(fn_name) {
-            FunctionKind::Core(func) => {
-                if func.deprecated() {
-                    exec_state.warn(CompilationError::err(
-                        self.callee.as_source_range(),
-                        format!("`{fn_name}` is deprecated, see the docs for a recommended replacement"),
-                    ));
-                }
-
-                let formals = func.args(false);
-
-                // If it's possible the input arg was meant to be labelled and we probably don't want to use
-                // it as the input arg, then treat it as labelled.
-                if let Some((Some(label), _)) = &args.kw_args.unlabeled {
-                    if (formals.iter().all(|a| a.label_required) || exec_state.pipe_value().is_some())
-                        && formals.iter().any(|a| &a.name == label && a.label_required)
-                        && !args.kw_args.labeled.contains_key(label)
-                    {
-                        let (label, arg) = args.kw_args.unlabeled.take().unwrap();
-                        args.kw_args.labeled.insert(label.unwrap(), arg);
-                    }
-                }
-
-                #[cfg(feature = "artifact-graph")]
-                let op = if func.feature_tree_operation() {
-                    let op_labeled_args = args
-                        .kw_args
-                        .labeled
-                        .iter()
-                        .map(|(k, arg)| (k.clone(), OpArg::new(OpKclValue::from(&arg.value), arg.source_range)))
-                        .collect();
-                    Some(Operation::StdLibCall {
-                        std_lib_fn: (&func).into(),
-                        unlabeled_arg: args
-                            .unlabeled_kw_arg_unconverted()
-                            .map(|arg| OpArg::new(OpKclValue::from(&arg.value), arg.source_range)),
-                        labeled_args: op_labeled_args,
-                        source_range: callsite,
-                        is_error: false,
-                    })
-                } else {
-                    None
-                };
-
-                for (label, arg) in &args.kw_args.labeled {
-                    match formals.iter().find(|p| &p.name == label) {
-                        Some(p) => {
-                            if !p.label_required {
-                                exec_state.err(CompilationError::err(
-                                    arg.source_range,
-                                    format!(
-                                        "The function `{fn_name}` expects an unlabeled first parameter (`{label}`), but it is labelled in the call"
-                                    ),
-                                ));
-                            }
-                        }
-                        None => {
-                            exec_state.err(CompilationError::err(
-                                arg.source_range,
-                                format!("`{label}` is not an argument of `{fn_name}`"),
-                            ));
-                        }
-                    }
-                }
-
-                // Attempt to call the function.
-                let mut return_value = {
-                    // Don't early-return in this block.
-                    exec_state.mut_stack().push_new_env_for_rust_call();
-                    let result = func.std_lib_fn()(exec_state, args).await;
-                    exec_state.mut_stack().pop_env();
-
-                    #[cfg(feature = "artifact-graph")]
-                    if let Some(mut op) = op {
-                        op.set_std_lib_call_is_error(result.is_err());
-                        // Track call operation.  We do this after the call
-                        // since things like patternTransform may call user code
-                        // before running, and we will likely want to use the
-                        // return value. The call takes ownership of the args,
-                        // so we need to build the op before the call.
-                        exec_state.global.operations.push(op);
-                    }
-
-                    result
-                }?;
-
-                update_memory_for_tags_of_geometry(&mut return_value, exec_state)?;
-
-                Ok(return_value)
-            }
-            FunctionKind::UserDefined => {
-                // Clone the function so that we can use a mutable reference to
-                // exec_state.
-                let func = fn_name.get_result(exec_state, ctx).await?.clone();
-
-                let Some(fn_src) = func.as_fn() else {
-                    return Err(KclError::Semantic(KclErrorDetails {
-                        message: "cannot call this because it isn't a function".to_string(),
-                        source_ranges: vec![callsite],
-                    }));
-                };
-
-                let return_value = fn_src
-                    .call_kw(Some(fn_name.to_string()), exec_state, ctx, args, callsite)
-                    .await
-                    .map_err(|e| {
-                        // Add the call expression to the source ranges.
-                        e.add_source_ranges(vec![callsite])
-                    })?;
-
-                let result = return_value.ok_or_else(move || {
-                    let mut source_ranges: Vec<SourceRange> = vec![callsite];
-                    // We want to send the source range of the original function.
-                    if let KclValue::Function { meta, .. } = func {
-                        source_ranges = meta.iter().map(|m| m.source_range).collect();
-                    };
-                    KclError::UndefinedValue(KclErrorDetails {
-                        message: format!("Result of user-defined function {} is undefined", fn_name),
-                        source_ranges,
-                    })
-                })?;
-
-                Ok(result)
-            }
-        }
-    }
-}
-
-fn update_memory_for_tags_of_geometry(result: &mut KclValue, exec_state: &mut ExecState) -> Result<(), KclError> {
-    // If the return result is a sketch or solid, we want to update the
-    // memory for the tags of the group.
-    // TODO: This could probably be done in a better way, but as of now this was my only idea
-    // and it works.
-    match result {
-        KclValue::Sketch { value } => {
-            for (name, tag) in value.tags.iter() {
-                if exec_state.stack().cur_frame_contains(name) {
-                    exec_state.mut_stack().update(name, |v, _| {
-                        v.as_mut_tag().unwrap().merge_info(tag);
-                    });
-                } else {
-                    exec_state
-                        .mut_stack()
-                        .add(
-                            name.to_owned(),
-                            KclValue::TagIdentifier(Box::new(tag.clone())),
-                            SourceRange::default(),
-                        )
-                        .unwrap();
-                }
-            }
-        }
-        KclValue::Solid { ref mut value } => {
-            for v in &value.value {
-                if let Some(tag) = v.get_tag() {
-                    // Get the past tag and update it.
-                    let tag_id = if let Some(t) = value.sketch.tags.get(&tag.name) {
-                        let mut t = t.clone();
-                        let Some(info) = t.get_cur_info() else {
-                            return Err(KclError::Internal(KclErrorDetails {
-                                message: format!("Tag {} does not have path info", tag.name),
-                                source_ranges: vec![tag.into()],
-                            }));
-                        };
-
-                        let mut info = info.clone();
-                        info.surface = Some(v.clone());
-                        info.sketch = value.id;
-                        t.info.push((exec_state.stack().current_epoch(), info));
-                        t
-                    } else {
-                        // It's probably a fillet or a chamfer.
-                        // Initialize it.
-                        TagIdentifier {
-                            value: tag.name.clone(),
-                            info: vec![(
-                                exec_state.stack().current_epoch(),
-                                TagEngineInfo {
-                                    id: v.get_id(),
-                                    surface: Some(v.clone()),
-                                    path: None,
-                                    sketch: value.id,
-                                },
-                            )],
-                            meta: vec![Metadata {
-                                source_range: tag.clone().into(),
-                            }],
-                        }
-                    };
-
-                    // update the sketch tags.
-                    value.sketch.merge_tags(Some(&tag_id).into_iter());
-
-                    if exec_state.stack().cur_frame_contains(&tag.name) {
-                        exec_state.mut_stack().update(&tag.name, |v, _| {
-                            v.as_mut_tag().unwrap().merge_info(&tag_id);
-                        });
-                    } else {
-                        exec_state
-                            .mut_stack()
-                            .add(
-                                tag.name.clone(),
-                                KclValue::TagIdentifier(Box::new(tag_id)),
-                                SourceRange::default(),
-                            )
-                            .unwrap();
-                    }
-                }
-            }
-
-            // Find the stale sketch in memory and update it.
-            if !value.sketch.tags.is_empty() {
-                let sketches_to_update: Vec<_> = exec_state
-                    .stack()
-                    .find_keys_in_current_env(|v| match v {
-                        KclValue::Sketch { value: sk } => sk.original_id == value.sketch.original_id,
-                        _ => false,
-                    })
-                    .cloned()
-                    .collect();
-
-                for k in sketches_to_update {
-                    exec_state.mut_stack().update(&k, |v, _| {
-                        let sketch = v.as_mut_sketch().unwrap();
-                        sketch.merge_tags(value.sketch.tags.values());
-                    });
-                }
-            }
-        }
-        KclValue::Tuple { value, .. } | KclValue::HomArray { value, .. } => {
-            for v in value {
-                update_memory_for_tags_of_geometry(v, exec_state)?;
-            }
-        }
-        _ => {}
-    }
-    Ok(())
-}
-
 impl Node<TagDeclarator> {
     pub async fn execute(&self, exec_state: &mut ExecState) -> Result<KclValue, KclError> {
         let memory_item = KclValue::TagIdentifier(Box::new(TagIdentifier {
@@ -1893,409 +1589,6 @@ impl Node<PipeExpression> {
     }
 }
 
-fn type_check_params_kw(
-    fn_name: Option<&str>,
-    function_expression: NodeRef<'_, FunctionExpression>,
-    args: &mut KwArgs,
-    exec_state: &mut ExecState,
-) -> Result<(), KclError> {
-    // If it's possible the input arg was meant to be labelled and we probably don't want to use
-    // it as the input arg, then treat it as labelled.
-    if let Some((Some(label), _)) = &args.unlabeled {
-        if (function_expression.params.iter().all(|p| p.labeled) || exec_state.pipe_value().is_some())
-            && function_expression
-                .params
-                .iter()
-                .any(|p| &p.identifier.name == label && p.labeled)
-            && !args.labeled.contains_key(label)
-        {
-            let (label, arg) = args.unlabeled.take().unwrap();
-            args.labeled.insert(label.unwrap(), arg);
-        }
-    }
-
-    for (label, arg) in &mut args.labeled {
-        match function_expression.params.iter().find(|p| &p.identifier.name == label) {
-            Some(p) => {
-                if !p.labeled {
-                    exec_state.err(CompilationError::err(
-                        arg.source_range,
-                        format!(
-                            "{} expects an unlabeled first parameter (`{label}`), but it is labelled in the call",
-                            fn_name
-                                .map(|n| format!("The function `{}`", n))
-                                .unwrap_or_else(|| "This function".to_owned()),
-                        ),
-                    ));
-                }
-
-                if let Some(ty) = &p.type_ {
-                    arg.value = arg
-                        .value
-                        .coerce(
-                            &RuntimeType::from_parsed(ty.inner.clone(), exec_state, arg.source_range).map_err(|e| KclError::Semantic(e.into()))?,
-                            exec_state,
-                        )
-                        .map_err(|e| {
-                            let mut message = format!(
-                                "{label} requires a value with type `{}`, but found {}",
-                                ty.inner,
-                                arg.value.human_friendly_type(),
-                            );
-                            if let Some(ty) = e.explicit_coercion {
-                                // TODO if we have access to the AST for the argument we could choose which example to suggest.
-                                message = format!("{message}\n\nYou may need to add information about the type of the argument, for example:\n  using a numeric suffix: `42{ty}`\n  or using type ascription: `foo(): number({ty})`");
-                            }
-                            KclError::Semantic(KclErrorDetails {
-                                message,
-                                source_ranges: vec![arg.source_range],
-                            })
-                        })?;
-                }
-            }
-            None => {
-                exec_state.err(CompilationError::err(
-                    arg.source_range,
-                    format!(
-                        "`{label}` is not an argument of {}",
-                        fn_name
-                            .map(|n| format!("`{}`", n))
-                            .unwrap_or_else(|| "this function".to_owned()),
-                    ),
-                ));
-            }
-        }
-    }
-
-    if !args.errors.is_empty() {
-        let actuals = args.labeled.keys();
-        let formals: Vec<_> = function_expression
-            .params
-            .iter()
-            .filter_map(|p| {
-                if !p.labeled {
-                    return None;
-                }
-
-                let name = &p.identifier.name;
-                if actuals.clone().any(|a| a == name) {
-                    return None;
-                }
-
-                Some(format!("`{name}`"))
-            })
-            .collect();
-
-        let suggestion = if formals.is_empty() {
-            String::new()
-        } else {
-            format!("; suggested labels: {}", formals.join(", "))
-        };
-
-        let mut errors = args.errors.iter().map(|e| {
-            CompilationError::err(
-                e.source_range,
-                format!("This argument needs a label, but it doesn't have one{suggestion}"),
-            )
-        });
-
-        let first = errors.next().unwrap();
-        errors.for_each(|e| exec_state.err(e));
-
-        return Err(KclError::Semantic(first.into()));
-    }
-
-    if let Some(arg) = &mut args.unlabeled {
-        if let Some(p) = function_expression.params.iter().find(|p| !p.labeled) {
-            if let Some(ty) = &p.type_ {
-                arg.1.value = arg
-                    .1
-                    .value
-                    .coerce(
-                        &RuntimeType::from_parsed(ty.inner.clone(), exec_state, arg.1.source_range)
-                            .map_err(|e| KclError::Semantic(e.into()))?,
-                        exec_state,
-                    )
-                    .map_err(|_| {
-                        KclError::Semantic(KclErrorDetails {
-                            message: format!(
-                                "The input argument of {} requires a value with type `{}`, but found {}",
-                                fn_name
-                                    .map(|n| format!("`{}`", n))
-                                    .unwrap_or_else(|| "this function".to_owned()),
-                                ty.inner,
-                                arg.1.value.human_friendly_type()
-                            ),
-                            source_ranges: vec![arg.1.source_range],
-                        })
-                    })?;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn assign_args_to_params_kw(
-    fn_name: Option<&str>,
-    function_expression: NodeRef<'_, FunctionExpression>,
-    mut args: Args,
-    exec_state: &mut ExecState,
-) -> Result<(), KclError> {
-    type_check_params_kw(fn_name, function_expression, &mut args.kw_args, exec_state)?;
-
-    // Add the arguments to the memory.  A new call frame should have already
-    // been created.
-    let source_ranges = vec![function_expression.into()];
-
-    for param in function_expression.params.iter() {
-        if param.labeled {
-            let arg = args.kw_args.labeled.get(&param.identifier.name);
-            let arg_val = match arg {
-                Some(arg) => arg.value.clone(),
-                None => match param.default_value {
-                    Some(ref default_val) => KclValue::from_default_param(default_val.clone(), exec_state),
-                    None => {
-                        return Err(KclError::Semantic(KclErrorDetails {
-                            source_ranges,
-                            message: format!(
-                                "This function requires a parameter {}, but you haven't passed it one.",
-                                param.identifier.name
-                            ),
-                        }));
-                    }
-                },
-            };
-            exec_state
-                .mut_stack()
-                .add(param.identifier.name.clone(), arg_val, (&param.identifier).into())?;
-        } else {
-            let unlabelled = args.unlabeled_kw_arg_unconverted();
-
-            let Some(unlabeled) = unlabelled else {
-                let param_name = &param.identifier.name;
-                return Err(if args.kw_args.labeled.contains_key(param_name) {
-                    KclError::Semantic(KclErrorDetails {
-                        source_ranges,
-                        message: format!("The function does declare a parameter named '{param_name}', but this parameter doesn't use a label. Try removing the `{param_name}:`"),
-                    })
-                } else {
-                    KclError::Semantic(KclErrorDetails {
-                        source_ranges,
-                        message: "This function expects an unlabeled first parameter, but you haven't passed it one."
-                            .to_owned(),
-                    })
-                });
-            };
-            exec_state.mut_stack().add(
-                param.identifier.name.clone(),
-                unlabeled.value.clone(),
-                (&param.identifier).into(),
-            )?;
-        }
-    }
-
-    Ok(())
-}
-
-fn coerce_result_type(
-    result: Result<Option<KclValue>, KclError>,
-    function_expression: NodeRef<'_, FunctionExpression>,
-    exec_state: &mut ExecState,
-) -> Result<Option<KclValue>, KclError> {
-    if let Ok(Some(val)) = result {
-        if let Some(ret_ty) = &function_expression.return_type {
-            let ty = RuntimeType::from_parsed(ret_ty.inner.clone(), exec_state, ret_ty.as_source_range())
-                .map_err(|e| KclError::Semantic(e.into()))?;
-            let val = val.coerce(&ty, exec_state).map_err(|_| {
-                KclError::Semantic(KclErrorDetails {
-                    message: format!(
-                        "This function requires its result to be of type `{}`, but found {}",
-                        ty.human_friendly_type(),
-                        val.human_friendly_type(),
-                    ),
-                    source_ranges: ret_ty.as_source_ranges(),
-                })
-            })?;
-            Ok(Some(val))
-        } else {
-            Ok(Some(val))
-        }
-    } else {
-        result
-    }
-}
-
-async fn call_user_defined_function_kw(
-    fn_name: Option<&str>,
-    args: Args,
-    memory: EnvironmentRef,
-    function_expression: NodeRef<'_, FunctionExpression>,
-    exec_state: &mut ExecState,
-    ctx: &ExecutorContext,
-) -> Result<Option<KclValue>, KclError> {
-    // Create a new environment to execute the function body in so that local
-    // variables shadow variables in the parent scope.  The new environment's
-    // parent should be the environment of the closure.
-    exec_state.mut_stack().push_new_env_for_call(memory);
-    if let Err(e) = assign_args_to_params_kw(fn_name, function_expression, args, exec_state) {
-        exec_state.mut_stack().pop_env();
-        return Err(e);
-    }
-
-    // Execute the function body using the memory we just created.
-    let result = ctx
-        .exec_block(&function_expression.body, exec_state, BodyType::Block)
-        .await;
-    let mut result = result.map(|_| {
-        exec_state
-            .stack()
-            .get(memory::RETURN_NAME, function_expression.as_source_range())
-            .ok()
-            .cloned()
-    });
-
-    result = coerce_result_type(result, function_expression, exec_state);
-
-    // Restore the previous memory.
-    exec_state.mut_stack().pop_env();
-
-    result
-}
-
-impl FunctionSource {
-    pub async fn call_kw(
-        &self,
-        fn_name: Option<String>,
-        exec_state: &mut ExecState,
-        ctx: &ExecutorContext,
-        mut args: Args,
-        callsite: SourceRange,
-    ) -> Result<Option<KclValue>, KclError> {
-        match self {
-            FunctionSource::Std { func, ast, props } => {
-                if props.deprecated {
-                    exec_state.warn(CompilationError::err(
-                        callsite,
-                        format!(
-                            "`{}` is deprecated, see the docs for a recommended replacement",
-                            props.name
-                        ),
-                    ));
-                }
-
-                type_check_params_kw(Some(&props.name), ast, &mut args.kw_args, exec_state)?;
-
-                if let Some(arg) = &mut args.kw_args.unlabeled {
-                    if let Some(p) = ast.params.iter().find(|p| !p.labeled) {
-                        if let Some(ty) = &p.type_ {
-                            arg.1.value = arg
-                                .1
-                                .value
-                                .coerce(
-                                    &RuntimeType::from_parsed(ty.inner.clone(), exec_state, arg.1.source_range)
-                                        .map_err(|e| KclError::Semantic(e.into()))?,
-                                    exec_state,
-                                )
-                                .map_err(|_| {
-                                    KclError::Semantic(KclErrorDetails {
-                                        message: format!(
-                                            "The input argument of {} requires a value with type `{}`, but found {}",
-                                            props.name,
-                                            ty.inner,
-                                            arg.1.value.human_friendly_type(),
-                                        ),
-                                        source_ranges: vec![callsite],
-                                    })
-                                })?;
-                        }
-                    }
-                }
-
-                #[cfg(feature = "artifact-graph")]
-                let op = if props.include_in_feature_tree {
-                    let op_labeled_args = args
-                        .kw_args
-                        .labeled
-                        .iter()
-                        .map(|(k, arg)| (k.clone(), OpArg::new(OpKclValue::from(&arg.value), arg.source_range)))
-                        .collect();
-                    Some(Operation::KclStdLibCall {
-                        name: fn_name.unwrap_or_default(),
-                        unlabeled_arg: args
-                            .unlabeled_kw_arg_unconverted()
-                            .map(|arg| OpArg::new(OpKclValue::from(&arg.value), arg.source_range)),
-                        labeled_args: op_labeled_args,
-                        source_range: callsite,
-                        is_error: false,
-                    })
-                } else {
-                    None
-                };
-
-                // Attempt to call the function.
-                exec_state.mut_stack().push_new_env_for_rust_call();
-                let mut result = {
-                    // Don't early-return in this block.
-                    let result = func(exec_state, args).await;
-                    exec_state.mut_stack().pop_env();
-
-                    #[cfg(feature = "artifact-graph")]
-                    if let Some(mut op) = op {
-                        op.set_std_lib_call_is_error(result.is_err());
-                        // Track call operation.  We do this after the call
-                        // since things like patternTransform may call user code
-                        // before running, and we will likely want to use the
-                        // return value. The call takes ownership of the args,
-                        // so we need to build the op before the call.
-                        exec_state.global.operations.push(op);
-                    }
-                    result
-                }?;
-
-                update_memory_for_tags_of_geometry(&mut result, exec_state)?;
-
-                Ok(Some(result))
-            }
-            FunctionSource::User { ast, memory, .. } => {
-                // Track call operation.
-                #[cfg(feature = "artifact-graph")]
-                {
-                    let op_labeled_args = args
-                        .kw_args
-                        .labeled
-                        .iter()
-                        .map(|(k, arg)| (k.clone(), OpArg::new(OpKclValue::from(&arg.value), arg.source_range)))
-                        .collect();
-                    exec_state.global.operations.push(Operation::GroupBegin {
-                        group: Group::FunctionCall {
-                            name: fn_name.clone(),
-                            function_source_range: ast.as_source_range(),
-                            unlabeled_arg: args
-                                .kw_args
-                                .unlabeled
-                                .as_ref()
-                                .map(|arg| OpArg::new(OpKclValue::from(&arg.1.value), arg.1.source_range)),
-                            labeled_args: op_labeled_args,
-                        },
-                        source_range: callsite,
-                    });
-                }
-
-                let result =
-                    call_user_defined_function_kw(fn_name.as_deref(), args, *memory, ast, exec_state, ctx).await;
-
-                // Track return operation.
-                #[cfg(feature = "artifact-graph")]
-                exec_state.global.operations.push(Operation::GroupEnd);
-
-                result
-            }
-            FunctionSource::None => unreachable!(),
-        }
-    }
-}
-
 #[cfg(test)]
 mod test {
     use std::sync::Arc;
@@ -2305,150 +1598,9 @@ mod test {
     use super::*;
     use crate::{
         exec::UnitType,
-        execution::{memory::Stack, parse_execute, ContextType},
-        parsing::ast::types::{DefaultParamVal, Identifier, Parameter},
+        execution::{parse_execute, ContextType},
         ExecutorSettings, UnitLen,
     };
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_assign_args_to_params() {
-        // Set up a little framework for this test.
-        fn mem(number: usize) -> KclValue {
-            KclValue::Number {
-                value: number as f64,
-                ty: NumericType::count(),
-                meta: Default::default(),
-            }
-        }
-        fn ident(s: &'static str) -> Node<Identifier> {
-            Node::no_src(Identifier {
-                name: s.to_owned(),
-                digest: None,
-            })
-        }
-        fn opt_param(s: &'static str) -> Parameter {
-            Parameter {
-                identifier: ident(s),
-                type_: None,
-                default_value: Some(DefaultParamVal::none()),
-                labeled: true,
-                digest: None,
-            }
-        }
-        fn req_param(s: &'static str) -> Parameter {
-            Parameter {
-                identifier: ident(s),
-                type_: None,
-                default_value: None,
-                labeled: true,
-                digest: None,
-            }
-        }
-        fn additional_program_memory(items: &[(String, KclValue)]) -> Stack {
-            let mut program_memory = Stack::new_for_tests();
-            for (name, item) in items {
-                program_memory
-                    .add(name.clone(), item.clone(), SourceRange::default())
-                    .unwrap();
-            }
-            program_memory
-        }
-        // Declare the test cases.
-        for (test_name, params, args, expected) in [
-            ("empty", Vec::new(), Vec::new(), Ok(additional_program_memory(&[]))),
-            (
-                "all params required, and all given, should be OK",
-                vec![req_param("x")],
-                vec![("x", mem(1))],
-                Ok(additional_program_memory(&[("x".to_owned(), mem(1))])),
-            ),
-            (
-                "all params required, none given, should error",
-                vec![req_param("x")],
-                vec![],
-                Err(KclError::Semantic(KclErrorDetails {
-                    source_ranges: vec![SourceRange::default()],
-                    message: "This function requires a parameter x, but you haven't passed it one.".to_owned(),
-                })),
-            ),
-            (
-                "all params optional, none given, should be OK",
-                vec![opt_param("x")],
-                vec![],
-                Ok(additional_program_memory(&[("x".to_owned(), KclValue::none())])),
-            ),
-            (
-                "mixed params, too few given",
-                vec![req_param("x"), opt_param("y")],
-                vec![],
-                Err(KclError::Semantic(KclErrorDetails {
-                    source_ranges: vec![SourceRange::default()],
-                    message: "This function requires a parameter x, but you haven't passed it one.".to_owned(),
-                })),
-            ),
-            (
-                "mixed params, minimum given, should be OK",
-                vec![req_param("x"), opt_param("y")],
-                vec![("x", mem(1))],
-                Ok(additional_program_memory(&[
-                    ("x".to_owned(), mem(1)),
-                    ("y".to_owned(), KclValue::none()),
-                ])),
-            ),
-            (
-                "mixed params, maximum given, should be OK",
-                vec![req_param("x"), opt_param("y")],
-                vec![("x", mem(1)), ("y", mem(2))],
-                Ok(additional_program_memory(&[
-                    ("x".to_owned(), mem(1)),
-                    ("y".to_owned(), mem(2)),
-                ])),
-            ),
-        ] {
-            // Run each test.
-            let func_expr = &Node::no_src(FunctionExpression {
-                params,
-                body: Program::empty(),
-                return_type: None,
-                digest: None,
-            });
-            let labeled = args
-                .iter()
-                .map(|(name, value)| {
-                    let arg = Arg::new(value.clone(), SourceRange::default());
-                    ((*name).to_owned(), arg)
-                })
-                .collect::<IndexMap<_, _>>();
-            let exec_ctxt = ExecutorContext {
-                engine: Arc::new(Box::new(
-                    crate::engine::conn_mock::EngineConnection::new().await.unwrap(),
-                )),
-                fs: Arc::new(crate::fs::FileManager::new()),
-                stdlib: Arc::new(crate::std::StdLib::new()),
-                settings: Default::default(),
-                context_type: ContextType::Mock,
-            };
-            let mut exec_state = ExecState::new(&exec_ctxt);
-            exec_state.mod_local.stack = Stack::new_for_tests();
-
-            let args = Args::new_kw(
-                KwArgs {
-                    unlabeled: None,
-                    labeled,
-                    errors: Vec::new(),
-                },
-                SourceRange::default(),
-                exec_ctxt,
-                None,
-            );
-            let actual =
-                assign_args_to_params_kw(None, func_expr, args, &mut exec_state).map(|_| exec_state.mod_local.stack);
-            assert_eq!(
-                actual, expected,
-                "failed test '{test_name}':\ngot {actual:?}\nbut expected\n{expected:?}"
-            );
-        }
-    }
 
     #[tokio::test(flavor = "multi_thread")]
     async fn ascription() {
